demos/01-essentials/01-ai-foundry/04-rag/rag-azure-py/pyproject.toml
[project]
name = "rag-azure"
version = "0.1.0"
description = "Azure AI Foundry RAG Demo"
readme = "readme.md"
requires-python = ">=3.9"
dependencies = [
    "azure-ai-projects",
    "azure-ai-inference",
    "azure-ai-evaluation",
    "azure-identity",
    "azure-search-documents",
    "python-dotenv",
    "azure-monitor-opentelemetry",
    "azure-storage-blob",
    "uvicorn[standard]",
    "tqdm",
]